Рефакторинг приложения. Разбивка репозиториев и перенос их в отдельный пакет
