+++ app/src/main/java/ru/marslab/filmoteca/domain/di/RepositoryModule.kt
@@ -5,6 +5,7 @@ import dagger.Provides
 import dagger.hilt.InstallIn
 import dagger.hilt.android.components.ViewModelComponent
 import dagger.hilt.android.scopes.ViewModelScoped
+import ru.marslab.filmoteca.data.GuestRepositoryImpl
 import ru.marslab.filmoteca.data.RepositoryNetworkImpl
 import ru.marslab.filmoteca.data.retrofit.MovieApi
 import ru.marslab.filmoteca.domain.repository.GuestRepository
@@ -19,5 +20,5 @@ object RepositoryModule {
 
     @ViewModelScoped
     @Provides
-    fun provideGuestRepository(api: MovieApi): GuestRepository = RepositoryNetworkImpl(api)
+    fun provideGuestRepository(api: MovieApi): GuestRepository = GuestRepositoryImpl(api)
 }
